Getting started features. - New section implemented for new user guidance for the first steps on the app. - Fixed: Reactivity on Stats component. - Minor styling optimmizations. - Fixed: API service to update account wasn't working due to missing header.

+++ src/services/accountAPI.js
@@ -1,4 +1,4 @@
 import API from './API';
 
-export const Update = ({ OTP, updateEntries }) => API.patch("/account", updateEntries, { headers: { OTP } });
-export const Delete = ({ OTP }) => API.delete('/account', { headers: { OTP } });
+export const Update = ({ OTP, updateEntries, token }) => API.patch("/account", updateEntries, { headers: { OTP, Authorization: `bearer ${token}` } });
+export const Delete = ({ OTP, token }) => API.delete('/account', { headers: { OTP, Authorization: `bearer ${token}` } });
